user 접근제어 기능 추가, 아직 CRUD.js 사용안함, topic.js connect error

+++ ExpressG/myapp/routes/topic.js
@@ -67,7 +67,6 @@ function deleteDatabase(id){
             console.log('connection good')
         }   
     });
-    console.log('함수 안에서 id 값',id)
     const deletequery = new Query(`
     DELETE FROM topics WHERE id='${id}'`);
     client.query(deletequery)
@@ -86,7 +85,7 @@ function updateDatabase(title, description, id){
     client.query(updatequery)
 }
 
-//create submit -> data directory에 저장
+//create process
 router.post('/create_process', (request, response) => {
     var post = request.body;
     var title = post.title;
@@ -121,20 +120,14 @@ router.get('/update/:pageId', (request, response) => {
         var description = topic.description;
         var list = template.list(res.rows);
         var html = template.HTML(title, list,
-            `
-            <form action="/topic/update_process"
+            `<form action="/topic/update_process"
             method="post">
                 <input type="hidden" name="id" value="${topic.id}">
+                <input type="hidden" name="users_id" value="${topic.users_id}">
                 <p><input type="text" name="title" placeholder="title" value="${title}"></p>
-                <p>
-                    <textarea name="description" placeholder="description">${description}</textarea>
-                </p>
-                <p>
-                    <input type="submit">
-                </p>
-            </form>  
-            
-            `,
+                <p><textarea name="description" placeholder="description">${description}</textarea></p>
+                <p><input type="submit"></p>
+            </form>`,
             `<a href="/topic/create">create</a>
             <a href="/topic/update/${topic.id}">update</a>`,
             login.authStatusUI(request ,response));
@@ -142,30 +135,34 @@ router.get('/update/:pageId', (request, response) => {
         }) 
 });
 
-//data directory read -> update 후 write
+//update process
 router.post('/update_process', (request, response) => {
     var post = request.body;
     var title = post.title;
     var description = post.description;
     var id = post.id;
+    var users_id = post.users_id;
+    if(users_id != request.user){
+        return response.redirect(`/topic/${id}`);
+    }
     updateDatabase(title, description, id);
     response.redirect(`/topic/${id}`);
               
 });
 
-//data directory read -> unliink
+//delete process
 router.post('/delete_process', (request, response) => {
     //login 접근 제어
     if(login.loginRequire(request, response) === false){return false;}
     var post = request.body;
     var id = post.id;
-    console.log('삭제할 topic의 id',id)
+    var users_id = post.users_id;
+    //user 접근제어
+    if(users_id != request.user){
+        return response.redirect(`/topic/${id}`);
+    }
     deleteDatabase(id);
-
-    //if(topic.users_id != )
-
     response.redirect('/');
- 
 });
 
 //Home else
@@ -189,12 +186,14 @@ router.get('/:pageId', (request, response, next) => {
         var sanitizeDescription = sanitizeHtml(topic.description, {allowedTags:['h1']});
         var list = template.list(res.rows);
         var html = template.HTML(sanitizeTitle, list,
-            `<h2>${sanitizeTitle}</h2>${sanitizeDescription}`,
+            `<h2>${sanitizeTitle}</h2>${sanitizeDescription}
+            <p>사용자 값 : ${topic.users_id}</p>`,
             `<a href="/topic/create/">글쓰기</a>
             <a href="/topic/update/${topic.id}">글수정</a>
-    
+
             <form action="/topic/delete_process" method="post">
                 <input type="hidden" name="id" value="${topic.id}">
+                <input type="hidden" name="users_id" value="${topic.users_id}">
                 <input type="submit" value="글삭제">
             </form>`,
             login.authStatusUI(request ,response));
